add logging.sh for use with origin automated testing
This script is intented to be run as part of the OpenShift jenkins
test automation.  It expects to be run on a devenv ami which already
has the origin source code and binary tar balls.  It expects to be
run via vagrant-openshift with `vagrant test-origin-aggregated-logging`
The logging.sh script is intended to be run from the hack/testing
directory where the other tests are found.