upgrade/testing: handle upgrade 4
add check that indices made it in successfully

exclude deployer pod from count:
remove component=deployer label as we need some way to not count it in
the search for running components as it seems to be a race whether it
comes back Running or not.

also, there's no need to check for image streams since they're not
required for deployment any longer.

work around weird race with dc_ops not getting set