deployer: simplify inputs
All parameters to the deployer template are now optional, as is the
logging-deployer secret. Parameters and secret contents can now be
accepted from three sources (which are checked in this order):
1. A configmap named logging-deployer
2. A secret named logging-deployer
3. Environment variables
Most deployer parameters are deprecated, but they will still work.
However the intent is that most parameters will come from the configmap.