use kibana instead of test app - refactor test scripts
Move a lot of duplicated code into deployer/scripts/util.sh and
use this code in the test scripts.
The dependency on the test application was causing problems with CI
when github.com could not be reached.  This completely eliminates the
use of the test app.  Instead, just use kibana for generating container
logs with a uuid.

(cherry picked from commit 9e434c3ddfa4ccbfe2d6b0b63da1589207547587)

Conflicts:
	deployer/scripts/util.sh
	hack/testing/logging.sh
	hack/testing/test-curator.sh
	hack/testing/test-datetime-future.sh
	hack/testing/test-es-copy.sh
	hack/testing/test-fluentd-forward.sh