use kibana instead of test app - refactor test scripts
Move a lot of duplicated code into deployer/scripts/util.sh and
use this code in the test scripts.
The dependency on the test application was causing problems with CI
when github.com could not be reached.  This completely eliminates the
use of the test app.  Instead, just use kibana for generating container
logs with a uuid.