CI testing support for ansible deployment
The primary purpose of this commit is to allow us to do CI testing
with the openshift-ansible based deployment of logging rather than
the old style logging deployer pod.  This assumes a devenv style
environment with OpenShift binaries pre-built from source, and
configured and run manually.  Then, the environment is configured to
run ansible on the local machine to run the openshift_logging role
to install and configure the logging components.
Since ansible doesn't create the logging-fluentd-template, the test script
has to create it and configure it, because several of the tests still
use this template.
There are a few other changes such as
* Default value for USE_JOURNAL is "" - have fluentd figure out if using
  journal or not
* Additional debugging for curator and fluentd failures
* Other minor test cleanups
* add back tests to make sure components are running
* default USE_JOURNAL to empty in ansible inventory