tool to use to check speed of logs to Fluentd and Elasticsearch from source
* Must be root or run with sudo: `$ sudo ./check-log-speed.sh`
* Must be on the host, not inside a pod
* Must be run with system:admin rights `oc login --username=system:admin`
* Will only check the first Elasticsearch found e.g. in a cluster of 3
  Elasticsearch nodes (`oc get pods -l component=es` returns 3), only
  the first one returned will be used.

```
$ sudo ./check-log-speed.sh
```

By default, this will not print anything if there are no differences.  If
you want to see all of the messages, including when there is no difference,
use `$ VERBOSE=1 sudo -E ./check-log-speed.sh`

```
Elasticsearch is 8 seconds behind /var/log/messages
```
This calculates the timestamp of the last record in `/var/log/messages`
and compares that to the timestamp of the last .operations.* index record
in Elasticsearch.  In this example, it means there is about an 8 second
lag between when the record was written to `/var/log/messages` and when
it was stored in Elasticsearch.
```
Elasticsearch is 7 seconds behind Fluentd for /var/log/messages
```
This calculates the timestamp of the last record in `/var/log/messages`
and compares that to the timestamp of the last record in `/var/log/messages`
read by Fluentd.  In this example, it means there is about a 7 second lag
between when Fluentd read the record and stored it in Elasticsearch.
```
Fluentd is 1 seconds behind /var/log/messages
```
This calculates the timestamp of the last record in `/var/log/messages`
and compares that to the timestamp of the last record in `/var/log/messages`
read by Fluentd.  In this example, it means there is about a 1 second lag
between when Fluentd read the record and when the record was written to
`/var/log/messages`.
```
Fluentd is 206 bytes of offset behind /var/log/messages
```
This is the difference between the file length of `/var/log/messages`
and the last file offset in `/var/log/messages` read by Fluentd, in bytes.
```
Fluentd is 1 records behind /var/log/messages
```
This is number of records that Fluentd has not yet read from
`/var/log/messages`.
```
Elasticsearch .operations. index is 10 seconds behind /var/log/containers/docker-registry-1-k5gs5_default_registry-ef6bdf268b256d33afa82ffe8f22c4193b63fa0eb55a1f239c30354dea8e0f6a.log
Elasticsearch .operations. index is 10 seconds behind Fluentd for /var/log/containers/docker-registry-1-k5gs5_default_registry-ef6bdf268b256d33afa82ffe8f22c4193b63fa0eb55a1f239c30354dea8e0f6a.log
```
Similar to the above.  In addition to `/var/log/messages`, the script
reads every file in `/var/log/containers/*.log` and skips the empty files.
```
Skipped 15 empty files of total 32
```
This tells you how many files were examined in total, and how many of those
were skipped because they were empty.

If there are no records in Elasticsearch for a particular file, you will get
a message like this:
```
Elasticsearch has no records for /var/log/containers/docker-registry-1-k5gs5_default_registry-ef6bdf268b256d33afa82ffe8f22c4193b63fa0eb55a1f239c30354dea8e0f6a.log
```

If Fluentd has not read any records from a file, then the Fluentd differences
reported will be from the beginning of the file.

Fluentd has only 1 position to keep track of in the journal, which it
stores as the journal cursor in `/var/log/journal.pos`.  Journal uses
microsecond resolution timestamps, so the times reported by the script
are in second.microsecond format.
```
Fluentd is 6.123456 seconds behind the journal
```
The script gets the timestamp of the last record in the journal, and the
timestamp of the current Fluentd journal cursor position, and compares
the two.  In this case, the last record read by Fluentd is 6.123456 seconds
older than the last record in the journal.
```
Fluentd is 5 records behind the journal
```
This means that there are 5 records in the journal that Fluentd has
not yet read.
```
Elasticsearch operations index is 9.543210 seconds behind the journal
```
This means the timestamp on the last record written to the .operations.*
indices in Elasticsearch is 9.543210 seconds behind the timestamp of the
last operations record written to the journal.
```
Elasticsearch project index is 7.343810 seconds behind the journal
```
This means the timestamp on the last record written to any of the
project.* indices in Elasticsearch is 7.343810 seconds behind the
timestamp of the last container record written to the journal.
```
Elasticsearch operations index is 9.543210 seconds behind Fluentd
```
This means the timestamp on the last record written to the .operations.*
indices in Elasticsearch is 9.543210 seconds behind the timestamp of the
last operations record read by Fluentd.
```
Elasticsearch project index is 7.343810 seconds behind Fluentd
```
This means the timestamp on the last record written to any of the
project.* indices in Elasticsearch is 7.343810 seconds behind the
timestamp of the last container record read by Fluentd.