Reformat cluster end-to-end test to use Bash helpers
Much of the work being done by the cluster end-to-end test can be
simplified by using the `os::cmd` suite of Bash helpers from Origin as
well as the `oc rollout status` command.

Of note -- in `test/cluster/rollout.sh` we are dropping the explicit
tests for ReplicationControllers and Pods as we can assume that the
OpenShift platform will function correctly at this point, and we are
not deploying any ReplicationControllers or Pods directly.

In `test/cluster/functionality.sh` we are getting the list of indices
in ElasticSearch by using the `/_cat/indices` REST endpoint, not by
looking through the pod log. We are also checking each Kibana pod, not
just one.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>