<drop> Make check-logs.go pass when no entries are found in ES
Historically, the check-logs.sh script was sliently failing as the
errors reported in the script were not exiting with a non-zero error
code. A previous commit made the script correctly emit `os.Exit(1)`,
however the fix for this test is pending, so this is a commit to easily
revert once the fix lands.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>