allow tests to run against released software with NOSETUP flag to logging.sh
If logging.sh is run with the NOSETUP argument, it will skip the set up
of OpenShift and logging components, and will assume they are already
present.
In addition, clean up and make consistent the use of the VERBOSE and CLUSTER
environment variables, and document the current logging.sh variables.