Change startup polling logic to hit APIs and check pod readiness
The current approach of looking through container logs to determine if
the ElasticSearch or Kibana node is a specific state is not only fragile
as there is no compatibility guarantee on the log messages but also not
the most useful as it requires that a cluster is set up to log to the
console, which is not default in production clusters.

Specifically, to query for Kibana and ElasticSearch readiness we should
be able to (in a perfect world) simply wait for the `oc rollout status`
check to succeed, as a full rollout of the `DeploymentConfig` will
require that the containers report a ready state. However, there are two
issues with that approach today:

 - Kibana does not have a readiness probe set
 - the OpenShift-Ansible installer creates `DeploymentConfig`s that set
   the desired replica count to 0, which always causes the rollout to
   succeed as there is no work to do. When these deployments are scaled
   in the future, the rollout is not updated

Both of these issues are being worked on, but in the interim we can
continue checking for Kibana and ElasticSearch readiness in this test
by issuing the same `curl` commands that the `ReadinessProbe` will. This
check can be relaxed as the other changes merge.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>