Check to see if the Elasticsearch pod is a master using the API
In order to check if the Elasticsearch pod that we are currently
investigating is a master node or not in the Elasticsearch cluster, we
use two API endpoints:

    /_cluster/state/master_node
    /_nodes/_local

The first will give use the Elasticsearch master node ID at
`.master_node`; the second will give us the current node ID as the first
key in `.nodes`. We simply check to see if these two IDs are or are not
identical.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>