Remove check for index recovery on master ES nodes
Today, we are checking that indices are recovered when the ES node we
exec into is the master for the ES cluster. However, there does not seem
to be a simple way to run this check using the API instead of looking
through the logs. As a failure to recover indices on startup will also
fail all of the following tests, we can remove this explicit check in
lieu of them and remove our last dependency on log munging.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>