Be more lenient when checking Pod readiness
When we manually check for Kibana or Elasticsearch rediness by executing
`curl` inside of the pods, it is possible for us to detect that the
services have become ready before the `ReadinessProbe` on the `Pod` has,
so if we are not lenient in expecting the `Pod` to be ready after we
detect it, we can race with the probe. If we are more lenient, this will
never flake out.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>