Ensure gathering admin token doesn't switch projects
When determining the admin user and token, we can accidentally end up
switching projects if the admin user and the test user are not in the
same context. By recording the current project before switching over and
returning to that project afterword, we ensure that we do not make this
mistake.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>