fix get_test_user_token; use openshift_logging_kibana_hostname
get_test_user_token always returns the admin credentials, so it always
works.  This makes it use the command line arguments. This also adds tests
to make sure a normal user cannot see a project it is not a member of, and
cannot see .operations index.
This also uses `openshift_logging_kibana_hostname` instead of
`openshift_hosted_logging_hostname` which is not supported by
playbooks/byo/openshift-cluster/openshift-logging.yml