add tests and script for multi-tenancy with multi-index /msearch
Add a script `hack/enable-kibana-msearch-access.sh` which can be
used to add permission for users to perform /_msearch queries of
Elasticsearch, in the manner of Kibana.
Added tests for multi-tenancy including testing for this script
to make sure it works.
This commit also changes Elasticsearch so that it has the ability
to pass in different plugins at build time for testing.