Force log messages to be treated as UTF-8
The various origin services log messages including the character µ,
which is encoded in UTF-8 as the 2-byte sequence 0xC2B5. Both the
in_tail and in_systemd inputs treat messages as ASCII-8BIT by default,
with only the in_tail plugin providing a way to override this. As a
result, if one of these messages (or any other message with UTF-8
characters) is read in and later needs to be converted to UTF-8 (e.g. as
part of being marshaled to JSON), the conversion will fail with an
UndefinedConversionError. By forcing the message to be treated as UTF-8,
we avoid this issue and other potential issues arising from the use of
ASCII-8BIT.