Fluentd: one output tag, one output plugin
When ops cluster is configured, application tags are retagged
into output_tag and operation tags are to output_ops_tag in
the filter-post-* phase.  This is to avoid switching a buffer
chunk to the next chunk due to the tag change

If ops cluster is not configured, all tags are retagged into
output_tag, which does not match output-operations.conf.

To implement the one output tag per output plugin, @OUTPUT label
is introduced.  To enable the one output tag feature, the label
has to be declared in /etc/fluent/fluent.conf.
1) In the ansible role, <label @OUTPUT> is added to fluentd.conf.j2
   for fluentd and fluent.conf for mux.  They are available in the
   clean install.
2) For the upgrade case, set existing configmap to fluentd_config_
   contents for fluentd and fluentd_mux_config_contents for mux;
   Then <label @OUTPUT> needs to be manually added at the appropriate
   place by the administrators.  This scenario must be documented.
3) In fluentd/run.sh, check @OUTPUT label in fluent.conf.   If it
   does not exist, the one output tag per plugin feature is not
   enabled.