fluentd - available disk space for the file buffer
In the calculation of available disk space for the file buffer, 25% of the found
space was used, which limitation is not necessary for pvc but is for hostmount.
This patch takes care of it.

(cherry picked from commit f14d1c7f84f2f55c6519c0d98c9d83d7e125cba9)