add json parser for mux clients
When running fluentd as a mux client, the k8s meta filter is omitted.
This filter also has a "side effect" of parsing the message field if
it contains a JSON value.  We need a small plugin to do this on the
mux client side.  The code is essentially copied from the k8s meta
filter plugin.

(cherry picked from commit ab3372c7a8db81119371dcc32203b4ab2c86d312)