Bug 1493188 - operations namespace logs not sent to .operations index
https://bugzilla.redhat.com/show_bug.cgi?id=1493188
The format of the CONTAINER_NAME field changed from
k8s_container-name.container-hash_ ...
to
k8s_container-name_ ....
which caused problems for the retag filter.  Since we don't care about
the value of container-hash in this context, just ignore it.

Also added some test to ensure that operations namespace logs end up
in es-ops in the .operations.* indices and nowhere else.