Fix references to cpu-limits and test environment
This patch is a required sibling to the openshift-ansible PR
https://github.com/openshift/openshift-ansible/pull/5748.

Without this patch landing first, that PR will not past its regression
tests, due to a behavior of the test environment provided by this repo
where it attempts to remove a CPU limit resource path, but fails when it
is not found.  That failure is now always the case with the changes in
PR https://github.com/openshift/openshift-ansible/pull/5748.

(cherry picked from commit 6b6fe9e696add4d30cd5528e7418a9412ea296d2)