Stream output of docker to prevent travis timeout
Originally, the output of 'docker build' was stored in a variable and only
after the image was built, sent to STDOUT. Travis has 10 minute kill period
when nothing goes to STDOUT which lead to premature test failures if building
of the image took longer (occasionally happened with Elasticsearch image).

Also adds a timestamp to each output line to better evaluate time
requirements of image builds.