remove dashes in ident values used for term searches to avoid test flakes
Searches like this are failing randomly and often:

    message_uuid=$( uuidgen )
    qs='{"query":{"term":{"systemd.u.SYSLOG_IDENTIFIER":"'"${message_uuid}"'"}}}'
    curl_es $espod /.operations.*/_search -X POST -d $qs

it seems as if Elasticsearch is analyzing the systemd.u.SYSLOG_IDENTIFIER
field which causes exact matches to fail, even though it is defined as
a non_analyzed field.
See https://www.elastic.co/guide/en/elasticsearch/reference/2.4/query-dsl-term-query.html

Until we can figure this out, remove the dashes from the uuidgen output
so that there will be no analysis of the string.