reduce test flakes with false hits for ops projects
Because of the problem where index templates can be loaded after the
data, we cannot rely on exact match searches (query term).  Instead,
do a possibly "fuzzy" search for namespace_name, then use jq to further
parse and query the output.
In order to make sure the code which processes operational and
"openshift-*" projects is working, create log messages for those
projects.