add bulk index rejection test
This tests the ability of fluentd to handle bulk index rejections.
The test starts up several curl processes which hammer Elasticsearch
with bulk index requests which causes fluentd to receive bulk rejections.
The methodology is to add some messages for both projects and operations
indices which should be rejected and queued and eventually resent
successfully.
The test will ensure that Elasticsearch is rejecting bulk requests and
fluentd is receiving bulk rejections.
The test has to run last because it will cause other CI tests to have
test flakes due to bulk index rejections, which somehow become
persistent after this test runs.