parse log field for mux minimal mode for json-file logs
When using MUX_CLIENT_MODE=minimal and json-file logs, we have to
parse the `log` field - mux only runs in journald mode so expects
the `MESSAGE` field to contain the JSON value to parse.
In the mux tests, make sure the `MESSAGE` field is populated with
the contents of the `log` field so that it will be parsed.
There appears to be a new behavior with the internal openshift
logging which is causing test flakes in the mux test - query the
`message` field only instead of `_all` (all fields) for the test
data.