fluentd: determine if openshift uses cri-o from nodeconfig
Parse `etc/origin/node/node-config.yaml` config file for
`kubeletArguments -> container-runtime-endpoint` if it contains `crio`.

Change generation of `in_tail` plugin config to don't try parsing as json
but as a regex instead if crio is identified.

The regex used is:
/^(?<time>.+) (?<stream>stdout|stderr)( (?<logtag>.))? (?<message>.*)$/

and is based on the parsing logic in cri-o for 1.8 and 1.9:
https://github.com/kubernetes-incubator/cri-o/blob/release-1.8/oci/oci.go#L363
https://github.com/kubernetes-incubator/cri-o/blob/release-1.9/oci/oci.go#L363

NOTE: the logtag was introduced in 1.9 so it is put as optional here, and currently
doesn't appear to be working well with Elasticsearch logs so this PR doesn't try to
append and reconstruct multiline logs.