fix mux-client-mode by tagging raw messages with .raw suffix
The code relies on the presence of the @timestamp field to indicate
that the record has already been processed into the common data model.
However, some applications actually use the @timestamp field, so we
need some other way to denote to mux that the message is a raw
message.
Change fluentd running in MUX_CLIENT_MODE=minimal to add a `.raw`
tag suffix to unprocessed log records before sending to mux so that
mux will know the record is not processed.