Bug 1588416 - should fetch log from /elasticsearch/persistent/logging-es/
As suggested by @Anping Li in the bug, logging-dump.sh is having a support
for the paths which contain "persistent" in addition to the original ones.
Note: this is an adjustment to the change made by pr/1108.

Also, suppressing the stderr output from "oc rsync", e.g., "WARNING: cannot
use rsync: rsync not available in container".  When rsync is not available
in a pod, "oc rsync" falls back to tar.