change rsyslog test to use viaq/rsyslog image
Use the viaq/rsyslog container image instead of the copr
repo packages for the rsyslog tests.

Fix the k8s_filename.rulebase to workaround
https://github.com/rsyslog/rsyslog/pull/2824

rsyslog 8.30 uses case insensitive variable names, so we have
to do a little hacking around the way we handle MESSAGE/message
in order to make sure the final record sent uses "message", and
in order to support older versions of rsyslog if necessary.

kube-* projects are operations projects