use standalone json parser instead of k8s meta filter
We already had the functionality in filter_k8s_meta_for_mux_client.rb
so that was renamed to filter_parse_json_field.rb to give us a
general purpose parser filter for json valued fields.  That filter
was also enhanced to look for either the `log` or the `MESSAGE`
field so that we could get rid of some of the `USE_JOURNAL`
related logic and simplify the mux pipeline cases.
add fluentd unit tests for json parsing