make sure journal rotates/deletes; fluentd check for open fds
Change the journald settings to ensure that journal rotates
and deletes files during the test.
Use loader and verify-loader to generate load and check for
missing/duplicate records.
Use /proc/$pid/fd to make sure fluentd does not have any
deleted files open.
Note there is still a bug in fluentd/systemd - there are still
missing and/or duplicate records, but the number should be
very small.  The primary purpose of this test is to ensure that
fluentd does not keep deleted FDs open when the journal rotates.