fluentd will detect json-file or crio logs automatically
This adds a new internal parser - json_or_crio - the way it
works is very simple - if the text begins with '{', assume it
is JSON formatted, otherwise, assume it is CRI-O formatted.
The assumption is that checking 1 character at the beginning of
the string is going to add negligible overhead to the already
expensive regexp or JSON parsing.
The nice thing is that it considerably simplifies our configuration.
We can even run on nodes where an upgrade from docker to cri-o
has occurred and there are mixed docker and cri-o log files.
There is a new config parameter for our in_tail plugins:
- `crio_format` - this is the regexp which matches the cri-o
log files - the default value is in the new parser plugin so
specifying this in the `in_tail` config is not necessary.