Support fluentd 1.2.6 using rh-ruby25 scl base image
Use the rh-ruby25 scl base image for fluentd.  This allows us
to use the latest fluentd 1.2.6 and the latest components.

This means we do not have to carry the patches for prometheus
and concat.  We can also use the log rotation built-in to
fluentd 1.x.  There is no longer a separate secure_forward
plugin - the built-in forward plugin has TLS and secret_key
support.  The latest prometheus has TLS support.

Some of our internal plugins like parser_json_or_crio had to
be ported to use the fluentd 1.x internal api.

Another big change is that there is a new format for the
fluent-plugin-system pos file - It is in JSON format and
must have a .json suffix.  The fluentd run.sh attempts to
convert the old file to the new format file.

The config file format is a little stricter - it is
apparently now an error to have a @type copy plugin
with no <store> sections.  You also cannot have a
rewrite_tag plugin with an empty pattern.

Also cleaned up the fluentd-forward test - it needed to
use the genid plugin when sending to two outputs.