allow setting flush_at_shutdown for migration
If the env. var. `FLUSH_AT_SHUTDOWN` is set to `"true"`, then when
fluentd is shutting down, it will attempt to flush all of its
file buffer records to Elasticsearch or forward destination.
This is needed in data migration or fluentd upgrade scenarios where
we do not want to or cannot convert the buffers to the newer version.

Steps:
1) oc set env ds/logging-fluentd FLUSH_AT_SHUTDOWN=true
2) oc label node -l logging-infra-fluentd=true --overwrite \
    logging-infra-fluentd=false
3) confirm that /var/lib/fluentd is empty - it may take a few
   minutes
4) oc set env ds/logging-fluentd FLUSH_AT_SHUTDOWN-
5) oc label node -l logging-infra-fluentd=false --overwrite \
    logging-infra-fluentd=true

(cherry picked from commit 0a6bcaf9726f9a16fc8ab679843dc58598525601)