add support for SSL_PARTIAL_CHAIN; update fluentd and components
Due to OpenShift 4.0 using an intermediate CA for the serviceaccount
ca.crt, we have to tell fluentd/ruby/openssl to allow the use
of a partial CA chain.  SSL_PARTIAL_CHAIN is true by default.
This also updates fluentd to 1.3.2 and updates the fluentd
components.
This also gets rid of some settings in the k8s meta plugin that
we should instead use the defaults for.