support for testing a 4.0 style operator deployment
This introduces a new container - logging-ci-test-runner - that
is built by the new CI job for the `openshift_install_src`
type for ci-operator.
The container source is in openshift/ci-operator/bild-images.
The file `Dockerfile` is used by regular CI.  The file
`Dockerfile.full` is used when manually running CI jobs when
developing them.
The logging-ci-test-runner is used in two different places
in CI for two different purposes. The first is run as the `test`
container in the CI job.  This `test` container runs in the
test namespace in the api.ci environment.  This `test` container
will deploy logging in the new remote cluster, then will launch
the logging-ci-test-runner container again in the new
`logging-ci-test-runner` pod which runs the actual CI tests
in the new remote cluster in the `openshift-logging` namespace.
The tests run in a pod, so they have to be converted not to
depend on running commands which are not applicable inside a pod
like the `systemctl` command.  There are many failing tests
currently which have been ignored in entrypoint.sh.