Secure ES metrics endpoint with auth proxy
ES uses self-signed CA, which isn't trusted by external components,
namely Prometheus. To tackle this, we're adding a side-car proxy
in front of ES, that speak TLS to the outside and its certs are
signed by the openshift signing service. Communication between
the proxy and ES happens within the same network namespace.