fluentd - vendor in all gems as source
All fluentd gem files are now vendored into the source tree
as their expanded source (`gem unpack`).  The files are under
`fluentd/vendored_gem_src/$gemname` (could not use `vendor`
because that is gitignore'd because it is used by ruby libs
we include).

There is a new script - `hack/update-fluentd-vendor-gems.sh`
developers should periodically run this script to update the
vendored source.  The file `fluentd/Gemfile` is used by the
script to specify fluentd and the plugins we use, and can also
be used to lock down versions if needed.  This script also
updates the `fluentd/sources` file so we can keep track of
the exact gem file that went into the source.

When the image is built, the source is copied to the image.
The script `fluentd/install-gems.sh` is run inside the image
to install the gems from source.  There are various patches
that are applied to the source and gemspec files, then the
gem files are reconstructed `gem build $gemname.gemspec`,
then finally `gem install` is used to install the gems.

This also vendors in jemalloc as source