Bug 1684048 - Newly introduced log-rotation in fluentd not working
https://bugzilla.redhat.com/show_bug.cgi?id=1684048
Cause: The files that implement the new log rotation
functionality were not being copied to the correct
fluentd directory.

Consequence: Fluentd was not using log rotation and its log
files were not being rotated.

Fix: Change the container build to inspect the fluentd
gem to find out where to install the files.

Result: The files that implement log rotation are
copied to the correct directory for fluentd to use.