rsyslog - support building from source
The rsyslog image is built in two stages.  The first stage
copies the source code from `rsyslog/vendored_src` and builds
the RPMs.  The second stage copies the runtime RPMs from the
builder and installs them.

There are a few changes to the imported source.  The `SourceN`
directives are removed from the spec files, and some changes
are made in order to build on both RHEL7 and RHEL8.

The script `hack/update-rsyslog-vendor-src.sh` is used to
update the rsyslog source.