better enable and disable of cluster-logging-operator
We cannot entirely rely on setting Unmanaged because we have
no guaranteed way of knowing exactly when the operator will
stop managing the resources.  So, instead, just shut off the
operator.

This also fixes a test flake with rsyslog - some strange
`make` weirdness with needing to set the dependencies in the
right order, that sometimes container ADD/COPY does not
preserve.

There are several improvements to the access control tests so
that if a test fails we get a lot more information about what
exactly failed and what the acls looked like at that time.

The `oc` command is wrapped so that we can extend the
connection/request timeout and add some debugging.